requirements.txt
altair==5.2.0
annotated-types==0.6.0
attrs==23.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.8
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
isodate==0.6.1
Jinja2==3.1.2
jira==3.5.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pendulum==2.1.2
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
preshed==3.0.9
protobuf==4.25.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
safetensors==0.4.1
salesforce==0.0.1
scipy==1.11.4
simple-salesforce==1.12.5
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
textblob==0.17.1
thinc==8.2.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
tornado==6.4
tqdm==4.66.1
transformers==4.36.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
wasabi==1.1.2
weasel==0.3.4
zeep==4.2.1
zipp==3.17.0

python-dotenv~=1.0.0